refactor: introduce LoadingErrorMessage component
